Restore main-bundle code-splitting around contributions routes
The contributions routes use UserContributionsTabs (a Zod schema) for
validateSearch, which TanStack Router needs eagerly when registering
the route tree. After #583 dissolved the contributions feature module,
that schema was inlined into -contributions.tsx alongside the heavy
UserContributions component. Routes' eager validateSearch import then
dragged the component file (and its 100+ KB of transitive deps:
my-markdown, error-label, drawer, dialog, dropdown, tabs, popover, …)
into the eager bundle, ballooning index-*.js by ~452 kB raw / ~141 kB
gzipped.

Move the Zod schema back into its own tiny co-located file
(-contributions-tabs.ts) so the route's eager imports stay light and
the heavy component file remains lazy.

Bundle before: 1498 kB raw / 454 kB gz
Bundle after:  1046 kB raw / 313 kB gz
