+++ src/routes/_user/learn/-contributions-tabs.ts
@@ -1,0 +1,7 @@
+import * as z from 'zod'
+
+export const UserContributionsTabs = z.object({
+	contributionsTab: z
+		.enum(['requests', 'phrases', 'playlists', 'answers', 'comments'])
+		.optional(),
+})
